Added Color changes for the states
Added a few methods in general options that will change the color of a
state when called on. You just need to pass it the state ID.

Still TODO is to change the color back but that wont take long. I also
defined new methods in Turing widgers. See the redraw_* methods...